add refactors

+++ blazor_full_project_ef_net8.0/DA12025/Services/UserService.cs
@@ -23,6 +23,11 @@ public class UserService : IUserService
     public void UpdateUser(UserDTO userToUpdate)
     {
         User? user = _userRepository.Get(u => u.Email == userToUpdate.Email);
+        if (user == null)
+        {
+            throw new ArgumentException("Cannot find the specified user");
+        }
+
         user.Name = userToUpdate.Name;
         user.LastName = userToUpdate.LastName;
         user.Email = userToUpdate.Email;
@@ -45,6 +50,11 @@ public class UserService : IUserService
     public void DeleteUser(string email)
     {
         User? userToDelete = _userRepository.Get(u => u.Email == email);
+        if (userToDelete == null)
+        {
+            throw new ArgumentException("Cannot find user with this email");
+        }
+
         _userRepository.Delete(userToDelete);
     }
